+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,29 +1,36 @@
-import {Component, DoCheck, OnInit} from '@angular/core';
+import {Component, DoCheck, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {CartService} from '../../services/cart.service';
 import {ProductModel} from '../../../products/models/product.model';
+import {CartProductModel} from '../../models/cart.model';
+import {Observable} from 'rxjs';
 
 @Component({
   selector: 'app-cart-list',
   templateUrl: './cart-list.component.html',
   styleUrls: ['./cart-list.component.less']
 })
-export class CartListComponent implements OnInit, DoCheck {
+export class CartListComponent {
+  @Input()
+  cartProducts: CartProductModel[];
 
-  products: ProductModel[];
-  totalPrice: number;
+  @Output()
+  delete: EventEmitter<number> = new EventEmitter<number>();
 
-  constructor(public cartService: CartService) { }
+  @Output()
+  changeCount: EventEmitter<{ id: number, count: number }> = new EventEmitter<{ id: number, count: number }>();
 
-  ngOnInit() {
+  constructor() { }
 
+  onDelete(id: number): void {
+    this.delete.emit(id);
   }
 
-  // TODO: Зачем этот хук, сделайте гетер и все будет обновляться
-  ngDoCheck() {
-    this.products = this.cartService.products;
-    this.totalPrice = this.cartService.getTotalPrice();
+  onChangeCount(count: number, id: number): void {
+    if (count === 0) {
+      this.onDelete(id);
+      return;
+    }
 
+    this.changeCount.emit({ id, count });
   }
-
-
 }