+++ src/app/products/components/product-list/product-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {ProductsService} from '../../services/products.service';
 import {ProductModel} from '../../models/product.model';
 import {CartService} from '../../../cart/services/cart.service';
+import {Observable} from 'rxjs';
 
 @Component({
   selector: 'app-product-list',
@@ -9,7 +10,7 @@ import {CartService} from '../../../cart/services/cart.service';
   styleUrls: ['./product-list.component.less']
 })
 export class ProductListComponent implements OnInit {
-  products: ProductModel[];
+  products$: Observable<ProductModel[]>;
 
   constructor(
     private productService: ProductsService,
@@ -17,11 +18,15 @@ export class ProductListComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-    this.products = this.productService.getProducts();
+    this.products$ = this.productService.getProducts$();
   }
 
 
   onAddProductToCart(product: ProductModel): void {
-    this.cartService.addToCart(product);
+    this.cartService.addProduct({
+      id: product.id,
+      name: product.name,
+      price: product.price
+    });
   }
 }