+++ src/app/products/services/products.service.ts
@@ -1,11 +1,8 @@
 import { Injectable } from '@angular/core';
-import {ProductModel} from '../models/product.model';
-
-const productList: Array<ProductModel> = [
-  new ProductModel(1, 'Book', true, 10),
-  new ProductModel(2, 'Excel', true, 20),
-  new ProductModel(3, 'Samsung', false, 30)
-];
+import {IProductModel, ProductModel} from '../models/product.model';
+import {Observable, of} from 'rxjs';
+import {products} from '../mocks/products.mock';
+import {map} from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -14,8 +11,12 @@ export class ProductsService {
 
   constructor() { }
 
-
-  getProducts() {
-    return productList;
+  getProducts$(): Observable<ProductModel[]> {
+    return of(products)
+      .pipe(
+        map((items: IProductModel[]) => {
+          return items.map(product => new ProductModel(product));
+        })
+      );
   }
 }